+++ src/client/java/com/bedsbutendgame/client/config/BedsButEndgameModMenu.java
@@ -1,0 +1,11 @@
+package com.bedsbutendgame.client.config;
+
+import com.terraformersmc.modmenu.api.ConfigScreenFactory;
+import com.terraformersmc.modmenu.api.ModMenuApi;
+
+public final class BedsButEndgameModMenu implements ModMenuApi {
+	@Override
+	public ConfigScreenFactory<?> getModConfigScreenFactory() {
+		return BbeConfigScreen::new;
+	}
+}